refactor: clean code of root app
create a new custome theme provider
create initial custom hook for auth
create the initial config file
